% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{bfh_theme_presentation}
\alias{bfh_theme_presentation}
\title{Presentation Theme (Larger Text)}
\usage{
bfh_theme_presentation(colors = BFH_COLORS)
}
\arguments{
\item{colors}{Color palette}
}
\description{
Presentation Theme (Larger Text)
}
