% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{bfh_theme_print}
\alias{bfh_theme_print}
\title{Print Theme (Optimized for Paper)}
\usage{
bfh_theme_print(colors = BFH_COLORS)
}
\arguments{
\item{colors}{Color palette}
}
\description{
Print Theme (Optimized for Paper)
}
