% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{BFHCHARTS_OPT_ALLOW_GLOBALENV_INJECT}
\alias{BFHCHARTS_OPT_ALLOW_GLOBALENV_INJECT}
\title{Option name: opt-out for globalenv inject_assets warning}
\format{
An object of class \code{character} of length 1.
}
\usage{
BFHCHARTS_OPT_ALLOW_GLOBALENV_INJECT
}
\description{
When \code{TRUE}, suppresses the warning emitted by \code{.validate_inject_assets()}
for functions defined in \code{.GlobalEnv} (development convenience).
}
\keyword{datasets}
