% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_pdf.R
\name{build_typst_content}
\alias{build_typst_content}
\title{Build Typst Document Content}
\usage{
build_typst_content(chart_image, metadata, spc_stats, template, template_file)
}
\arguments{
\item{chart_image}{Path to chart PNG (can be absolute or relative)}

\item{metadata}{Metadata list}

\item{spc_stats}{SPC statistics list}

\item{template}{Template name}

\item{template_file}{Relative path to template .typ file (relative to document location)}
}
\value{
Character vector with Typst content
}
\description{
Build Typst Document Content
}
\keyword{internal}
