% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_types.R
\docType{data}
\name{CHART_TYPES_EN}
\alias{CHART_TYPES_EN}
\title{English Chart Type Codes}
\format{
Character vector of valid chart codes
}
\usage{
CHART_TYPES_EN
}
\description{
Valid qicharts2 chart type codes.
}
\keyword{datasets}
