% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfh_spc_analysis_class.R
\name{format.bfh_spc_analysis}
\alias{format.bfh_spc_analysis}
\title{Format Method for bfh_spc_analysis}
\usage{
\method{format}{bfh_spc_analysis}(x, ...)
}
\arguments{
\item{x}{A \code{bfh_spc_analysis} object.}

\item{...}{Ignored.}
}
\value{
Character scalar.
}
\description{
Returns a single-line character summary. Used by \code{paste()}, format-
contexts, and debugging-helpers.
}
