% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{format_time_value}
\alias{format_time_value}
\title{Format Time Values for Display}
\usage{
format_time_value(value, unit = "minutes")
}
\arguments{
\item{value}{Numeric time value}

\item{unit}{Unit of time ("minutes", "hours", "days")}
}
\value{
Formatted time string
}
\description{
Converts numeric time values (in minutes or hours) to readable format.
}
\keyword{internal}
