% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_y_axis_formatting.R
\name{format_unscaled_number}
\alias{format_unscaled_number}
\title{Format Unscaled Number with Danish Notation}
\usage{
format_unscaled_number(val)
}
\arguments{
\item{val}{Numeric value}
}
\value{
Formatted string
}
\description{
Format Unscaled Number with Danish Notation
}
\keyword{internal}
