% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_types.R
\name{get_chart_description}
\alias{get_chart_description}
\title{Get Chart Type Description}
\usage{
get_chart_description(chart_type)
}
\arguments{
\item{chart_type}{Chart type (engelsk kode eller dansk navn)}
}
\value{
Dansk beskrivelse af chart type
}
\description{
Henter dansk beskrivelse for et givent chart type.
}
\examples{
\dontrun{
get_chart_description("run")
get_chart_description("I-kort (Individuelle værdier)")
}
}
\seealso{
\link{CHART_TYPE_DESCRIPTIONS}, \code{\link[=get_qic_chart_type]{get_qic_chart_type()}}

Other spc-chart-types: 
\code{\link{CHART_TYPES_DA}},
\code{\link{CHART_TYPES_EN}},
\code{\link{CHART_TYPE_DESCRIPTIONS}},
\code{\link{chart_type_requires_denominator}()},
\code{\link{get_qic_chart_type}()}
}
\concept{spc-chart-types}
\keyword{internal}
