% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_date_formatting.R
\name{get_optimal_formatting}
\alias{get_optimal_formatting}
\title{Get Optimal Date Formatting Configuration}
\usage{
get_optimal_formatting(interval_info, debug = FALSE)
}
\arguments{
\item{interval_info}{List from \code{\link[=detect_date_interval]{detect_date_interval()}}}

\item{debug}{Logical, enable debug output (default: FALSE)}
}
\value{
List with formatting configuration:
\describe{
\item{labels}{Date format string or scales label function}
\item{breaks}{Break specification for scale}
\item{n_breaks}{Number of breaks}
\item{use_smart_labels}{Logical, use scales::label_date_short()}
}
}
\description{
Returns optimal x-axis formatting based on detected date intervals
and number of observations.
}
\examples{
dates <- seq(as.Date("2024-01-01"), by = "month", length.out = 24)
interval_info <- detect_date_interval(dates)
get_optimal_formatting(interval_info)
}
