% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{\%||\%}
\alias{\%||\%}
\title{Null-coalescing Operator}
\usage{
x \%||\% y
}
\arguments{
\item{x}{First value}

\item{y}{Second value (returned if x is NULL)}
}
\value{
x if not NULL, otherwise y
}
\description{
Returns the first non-NULL value.
}
\examples{
NULL \%||\% "default"  # Returns "default"
"value" \%||\% "default"  # Returns "value"
}
\seealso{
\code{\link[=get_y_axis_unit_label]{get_y_axis_unit_label()}}

Other spc-helpers: 
\code{\link{Y_AXIS_UNITS_DA}},
\code{\link{get_y_axis_unit_label}()}
}
\concept{spc-helpers}
