% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfh_spc_analysis_class.R
\name{is_bfh_spc_analysis}
\alias{is_bfh_spc_analysis}
\title{Check if Object is bfh_spc_analysis}
\usage{
is_bfh_spc_analysis(x)
}
\arguments{
\item{x}{Object to test.}
}
\value{
Logical indicating whether x inherits from
\code{bfh_spc_analysis}.
}
\description{
Check if Object is bfh_spc_analysis
}
