% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_date_formatting.R
\name{parse_danish_dates}
\alias{parse_danish_dates}
\title{Parse Danish Date Formats}
\usage{
parse_danish_dates(date_strings)
}
\arguments{
\item{date_strings}{Character vector of date strings}
}
\value{
POSIXct vector of parsed dates (NA for failed parses)
}
\description{
Attempts to parse dates in common Danish formats (dd-mm-yyyy).
}
\examples{
parse_danish_dates(c("01-01-2024", "15-03-2024", "31-12-2024"))
}
