% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfh_qic_result.R
\name{plot.bfh_qic_result}
\alias{plot.bfh_qic_result}
\title{Plot Method for bfh_qic_result}
\usage{
\method{plot}{bfh_qic_result}(x, ...)
}
\arguments{
\item{x}{A \code{bfh_qic_result} object}

\item{...}{Additional arguments passed to print.ggplot}
}
\value{
The ggplot object invisibly
}
\description{
Extracts and displays the ggplot object from a bfh_qic_result.
Enables use of generic plot() function.
}
\keyword{internal}
