% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_objects.R
\name{print.spc_plot_config}
\alias{print.spc_plot_config}
\title{Print SPC Plot Configuration}
\usage{
\method{print}{spc_plot_config}(x, ...)
}
\arguments{
\item{x}{SPC plot configuration object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns x
}
\description{
Print SPC Plot Configuration
}
\seealso{
\code{\link[=spc_plot_config]{spc_plot_config()}}

Other spc-config: 
\code{\link{FONT_SCALING_CONFIG}},
\code{\link{calculate_base_size}()},
\code{\link{phase_config}()},
\code{\link{print.phase_config}()},
\code{\link{print.viewport_dims}()},
\code{\link{spc_plot_config}()},
\code{\link{viewport_dims}()}
}
\concept{spc-config}
