% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_objects.R
\name{print.viewport_dims}
\alias{print.viewport_dims}
\title{Print Viewport Dimensions}
\usage{
\method{print}{viewport_dims}(x, ...)
}
\arguments{
\item{x}{Viewport dimensions object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns x
}
\description{
Print Viewport Dimensions
}
\seealso{
\code{\link[=viewport_dims]{viewport_dims()}}

Other spc-config: 
\code{\link{FONT_SCALING_CONFIG}},
\code{\link{calculate_base_size}()},
\code{\link{phase_config}()},
\code{\link{print.phase_config}()},
\code{\link{print.spc_plot_config}()},
\code{\link{spc_plot_config}()},
\code{\link{viewport_dims}()}
}
\concept{spc-config}
