% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{spc_themes}
\alias{spc_themes}
\title{SPC Theme Utilities}
\description{
Theme-related utilities specific to SPC charts. General theming is handled
by the BFHtheme package.
}
\seealso{
\code{\link[BFHtheme:theme_bfh]{BFHtheme::theme_bfh()}} for general BFH theming
}
