% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{themes}
\alias{themes}
\title{Hospital Branding and Themes}
\description{
Configurable theming system for SPC charts with multi-organizational support.
Default branding is BFH (Bispebjerg og Frederiksberg Hospital), but can be
customized for any healthcare organization.
}
