% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chart_types.R
\name{get_chart_description}
\alias{get_chart_description}
\title{Get Chart Type Description}
\usage{
get_chart_description(chart_type)
}
\arguments{
\item{chart_type}{Chart type (engelsk kode eller dansk navn)}
}
\value{
Dansk beskrivelse af chart type
}
\description{
Henter dansk beskrivelse for et givent chart type.
}
\examples{
get_chart_description("run")
get_chart_description("I-kort (Individuelle værdier)")
}
