% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bfh_qic_result.R
\name{is_bfh_qic_result}
\alias{is_bfh_qic_result}
\title{Check if Object is bfh_qic_result}
\usage{
is_bfh_qic_result(x)
}
\arguments{
\item{x}{Object to test}
}
\value{
Logical indicating whether x is a bfh_qic_result object
}
\description{
Check if Object is bfh_qic_result
}
\keyword{internal}
