% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{LABEL_PLACEMENT_COINCIDENT_THRESHOLD_FACTOR}
\alias{LABEL_PLACEMENT_COINCIDENT_THRESHOLD_FACTOR}
\title{Coincident-lines threshold factor}
\format{
An object of class \code{numeric} of length 1.
}
\usage{
LABEL_PLACEMENT_COINCIDENT_THRESHOLD_FACTOR
}
\description{
When \code{abs(yA_npc - yB_npc) < label_height_npc * THIS_FACTOR}, lines are
treated as effectively coincident; labels are placed one above and one
below the same line position.
}
\keyword{datasets}
