% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{MIN_BASELINE_N}
\alias{MIN_BASELINE_N}
\title{Minimum anbefalede baseline-observationer for stabile SPC-kontrolgrænser}
\format{
An object of class \code{integer} of length 1.
}
\usage{
MIN_BASELINE_N
}
\description{
Anhøj-reglerne og SPC-litteraturen (Anhøj & Olesen 2014) kræver ca. 8+
punkter for meningsfulde kontrolgrænser og pålidelig signal-detektion.
Under denne grænse er kontrolgrænserne statistisk usikre.
}
\keyword{datasets}
