% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config_objects.R
\name{print.viewport_dims}
\alias{print.viewport_dims}
\title{Print Viewport Dimensions}
\usage{
\method{print}{viewport_dims}(x, ...)
}
\arguments{
\item{x}{Viewport dimensions object}

\item{...}{Additional arguments (ignored)}
}
\value{
Invisibly returns x
}
\description{
Print Viewport Dimensions
}
