% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\docType{data}
\name{Y_AXIS_UNITS_DA}
\alias{Y_AXIS_UNITS_DA}
\title{Y-Axis Unit Labels (Danish)}
\format{
Named character vector
}
\usage{
Y_AXIS_UNITS_DA
}
\description{
Mapping of unit codes to Danish labels.
}
\seealso{
\code{\link[=get_y_axis_unit_label]{get_y_axis_unit_label()}}, \code{\link[=create_spc_chart]{create_spc_chart()}}

Other spc-helpers: 
\code{\link{get_y_axis_unit_label}()}
}
\concept{spc-helpers}
\keyword{datasets}
